# Optimize graph searcher