fixes #3048