@quic-sendilk 2024-06-13
This PR contains changes to optimize graph sercher logic.
The original method of graph searching involved traversing the entire graph starting at each input op, with a certain sliding window length determined by the longest pattern to match. At every new set of ops which the sliding window covers, all known patterns are checked. This leads to many unnecessary windows being checked and also does not scale well for larger pattern sizes. If the graph has multiple input ops, the entire search is repeated for each input op, further adding onto the time spent.

The proposed change instead identifies ops to start searching at for each pattern, and traverses down the graph only up to the pattern's length. This greatly reduces the number of ops being checked.
To be more specific, assume that [Conv, Bn] is a pattern. The new method will first identify all Conv ops in the connected graph, and start searching for [Conv, Bn] pattern only at those Conv ops. Once a Conv op and its child has been traversed, the pattern matching stops, and no ops after the Conv's child is traversed.

When profiling with 32 layer bertcache Llama MHA model, I saw the following:

Without pyinstrument:
old method: 20 seconds for qsim init, 5.85 seconds in graph search
new method: 14 seconds for qsim init, .0028 seconds in graph search

With pyinstrument:
old method: 1746 seconds for qsim init, 1725 seconds in graph search
new method: 20.8 seconds for qsim init, .006 seconds in graph search